Now with 24% more magic.

+++ autoupdate.ps1
@@ -1,3 +1,17 @@
+# define latest version logic
+$Headers = @{'Accept'='application/vnd.github.v3+json'}
+$Response = ( Invoke-WebRequest "https://api.github.com/repos/RocketChat/Rocket.Chat.Electron/releases/latest" -Method GET -Headers $Headers)
+If ( $Response -eq "200") {
+    
+}
+Else {
+    Write-Error "The API call didn't work as expected!"
+    Pause
+    Exit
+}
+
+
+
 ################################################
 # Get new version
 $Package = "rocketchat"
@@ -6,39 +20,52 @@ $VERSION = Read-Host "What version do you want to build?" #"3.0.3"
 # Populate Version # Get-Content? Scrape page?
 $TEMPDIR = ".\versions\template\"
 $WORKDIR = ".\versions\$VERSION\"
-# Get the Hash from the download
-$File = "C:\Temp\rocketchat-setup-$VERSION.exe"
-$FileURL = "https://github.com/RocketChat/Rocket.Chat.Electron/releases/download/$($VERSION)/rocketchat-setup-$($VERSION).exe"
-( New-Object System.Net.WebClient ).DownloadFile($FileURL,$File)
-$Hash = ( Get-FileHash $File -Algorithm SHA256 ).Hash
-Remove-Item $File
-# Make version folder
-New-Item -Path $WORKDIR -ItemType Directory
-New-Item -Path "$WORKDIR\tools" -ItemType Directory
-# Get nuspec template content, write to new file
-$Content = Get-ChildItem -Path "$TEMPDIR\rocketchat.nuspec" | Get-Content -Raw
-$Content -Replace '#VERSION#',$VERSION | Out-File "$WORKDIR\rocketchat.nuspec"
-# Same for install script
-$Content = Get-ChildItem -Path "$TEMPDIR\tools\chocolateyinstall.ps1" | Get-Content -Raw
-$Content = ( $Content -Replace '#VERSION#',$VERSION )
-$Content -Replace '#HASH#',$HASH | Out-File "$WORKDIR\tools\chocolateyinstall.ps1"
 
-################################################
-# Copy to Rocketchat folder
-################################################
-$NEWDIR = ".\rocketchat" 
-Remove-Item $NEWDIR -Recurse -Force
-New-Item -Path $NEWDIR -ItemType Directory
-
-Foreach ( $Item in ( Get-ChildItem $WORKDIR ) ) {
-    Copy-Item $Item.Fullname -Destination $NEWDIR -Recurse -Force
+If ( !( Test-Path "$WORKDIR\*.nuspec"  ) ) {
+    # Get the Hash from the download
+    $File = "C:\Temp\rocketchat-setup-$VERSION.exe"
+    $FileURL = "https://github.com/RocketChat/Rocket.Chat.Electron/releases/download/$($VERSION)/rocketchat-setup-$($VERSION).exe"
+    ( New-Object System.Net.WebClient ).DownloadFile($FileURL,$File)
+    $Hash = ( Get-FileHash $File -Algorithm SHA256 ).Hash
+    Remove-Item $File
+    # Make version folder
+    New-Item -Path $WORKDIR -ItemType Directory
+    New-Item -Path "$WORKDIR\tools" -ItemType Directory
+    # Get nuspec template content, write to new file
+    $Content = Get-ChildItem -Path "$TEMPDIR\rocketchat.nuspec" | Get-Content -Raw
+    $Content -Replace '#VERSION#',$VERSION | Out-File "$WORKDIR\rocketchat.nuspec"
+    # Same for install script
+    $Content = Get-ChildItem -Path "$TEMPDIR\tools\chocolateyinstall.ps1" | Get-Content -Raw
+    $Content = ( $Content -Replace '#VERSION#',$VERSION )
+    $Content -Replace '#HASH#',$HASH | Out-File "$WORKDIR\tools\chocolateyinstall.ps1"
 }
+
+#################################################
+## Copy to Rocketchat folder
+#################################################
+#$NEWDIR = ".\rocketchat" 
+#Remove-Item $NEWDIR -Recurse -Force 
+#New-Item -Path $NEWDIR -ItemType Directory
+#
+#Foreach ( $Item in ( Get-ChildItem $WORKDIR ) ) {
+#    Copy-Item $Item.Fullname -Destination $NEWDIR -Recurse -Force
+#}
+#
+
+# Add to git and push
+CD $WORKDIR 
+git add .
+CD ..\
+CD ..\
+git commit -m "Updated to version $VERSION"
+git push
+
 
 ################################################
 # Build & push to chocolatey
 ################################################
 $PKGFile = "C:\Temp\"
-$Nuspec = ( Get-ChildItem $NEWDIR\*.nuspec )
+$Nuspec = ( Get-ChildItem $WORKDIR\*.nuspec )
 choco pack $Nuspec.Fullname --output-directory=$PKGFile 
 Rename-Item ( Get-ChildItem "$PKGFile\$Package.$VERSION.nupkg" ) "$Package.nupkg"
 choco push ( Get-ChildItem "$PKGFile\$Package.nupkg" ).Fullname -s="https://chocolatey.org/"
